move utils/images to utils/image-quote
