+++ src/utils/image-quote/style-bnw.ts
